services/game-allocator-service/pyproject.toml
[project]
name = "game-allocator-service"
version = "0.1.0"
description = "Game allocator microservice for load balancing"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "aiohttp>=3.10.0",
    "asyncio>=4.0.0",
    "nats-py>=2.10.0",
    "numpy>=2.3.0",
    "prometheus-api-client>=0.6.0",
    "py-consul>=1.6.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.9.1",
    "python-json-logger>=3.3.0",
    "redis>=6.0.0",
]
[dependency-groups]
dev=[
    "ty>=0.0.16",
    "ruff>=0.15.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "fakeredis>=2.20.0",
    "pytest-cov>=4.1.0",
]